# packages/protocol/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "additionalProperties": false,
        "description": "One stable, machine-readable error.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ApiError",
        "type": "object"
      },
      "ApiErrorResponse": {
        "additionalProperties": false,
        "description": "The envelope every v1 failure is serialized into.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
      "EventPage": {
        "additionalProperties": false,
        "properties": {
          "earliest_available_sequence": {
            "title": "Earliest Available Sequence",
            "type": "integer"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/StreamEventEnvelope"
            },
            "title": "Items",
            "type": "array"
          },
          "latest_sequence": {
            "title": "Latest Sequence",
            "type": "integer"
          },
          "next_after_sequence": {
            "title": "Next After Sequence",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "next_after_sequence",
          "has_more",
          "latest_sequence",
          "earliest_available_sequence"
        ],
        "title": "EventPage",
        "type": "object"
      },
...
      "JsonValue": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "array"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "RuntimeEventType": {
        "enum": [
          "project.created",
          "project.updated",
          "project.deleted",
          "approval.requested",
          "approval.approved",
          "approval.rejected",
          "approval.consumed",
          "changeset.created",
          "changeset.status_changed",
          "changeset.rolled_back",
          "docker_action.requested",
          "docker_action.started",
          "docker_action.completed",
          "docker_action.failed",
          "remote_session.connected",
          "remote_session.disconnected",
          "remote_session.failed",
          "remote_operation.started",
          "remote_operation.completed",
          "log.subscription_started",
          "log.subscription_paused",
          "log.subscription_resumed",
          "log.subscription_deleted",
          "log.batch_written",
          "log.source_rotated",
          "log.backpressure",
          "log.subscription_error",
          "investigation.created",
          "investigation.started",
          "investigation.status_changed",
          "investigation.completed",
          "investigation.cancelled",
          "investigation.failed",
          "operation.queued",
          "operation.running",
          "operation.cancel_requested",
          "operation.succeeded",
          "operation.failed",
          "operation.cancelled",
          "operation.uncertain",
          "agent_run.started",
          "agent_run.status_changed",
          "agent_run.completed",
          "agent_run.failed",
          "agent_run.cancelled",
          "agent_hook",
          "agent.text.delta",
          "agent.message.completed",
          "model_round.started",
          "model_round.completed",
          "tool.proposed",
          "policy.decided",
          "todo.changed",
          "hypothesis.changed",
          "conclusion.created",
          "context.compacted",
          "safety_state.changed",
          "tool_call.started",
          "tool_call.status_changed",
          "tool_call.completed",
          "child_run.started",
          "child_run.completed",
          "evidence.appended",
          "registry_proposal.created",
          "registry_proposal.decided",
          "recovery.started",
          "recovery.completed"
        ],
        "title": "RuntimeEventType",
        "type": "string"
      },
...
      "StreamEventEnvelope": {
        "additionalProperties": false,
        "properties": {
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "event_type": {
            "anyOf": [
              {
                "enum": [
                  "stream.hello",
                  "stream.heartbeat",
                  "stream.gap",
                  "stream.slow_consumer"
                ],
                "type": "string"
              },
              {
                "$ref": "#/components/schemas/RuntimeEventType"
              }
            ],
            "title": "Event Type"
          },
          "investigation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investigation Id"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "schema_version": {
            "const": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "sequence": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sequence"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          }
        },
        "required": [
          "schema_version",
          "event_type",
          "occurred_at"
        ],
        "title": "StreamEventEnvelope",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "IncidentLens Product API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "operationId": "listProductEvents",
        "parameters": [
          {
            "in": "query",
            "name": "after_sequence",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Sequence",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "in": "query",
            "name": "target_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Id"
            }
          },
          {
            "in": "query",
            "name": "investigation_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Investigation Id"
            }
          },
          {
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Invalid event cursor or filter"
          }
        },
        "summary": "List Product Events",
        "tags": [
          "events"
        ]
      }
    },
...
  }
}